# swagger/chainquery.yaml
tags:
  -
    name: stat
    description: 'Operations that describe the status of ChainQuery'
  -
    name: query
    description: 'Operations dealing with raw queries to the chainquery database.'
paths:
  /Status:
    get:
      operationId: status
      summary: 'Returns important status information about Chain Query'
      consumes:
        - application/json
      produces:
        - application/json
      schemes:
        - http
      tags:
        - stat
      responses:
        '200':
          description: 'Chain Query status'
          schema:
            x-oad-type: reference
            $ref: '#/definitions/TableStatus'
          x-oad-type: response
    x-oad-type: operation
  /AddressSummary:
    get:
      operationId: AddressSummary
      summary: 'Returns a summary of Address activity'
      description: 'It returns sent, recieved, balance, and number of transactions it has been used in.'
      consumes:
        - application/json
      produces:
        - application/json
      schemes:
        - http
      tags:
        - stat
      responses:
        '200':
          description: 'It returns sent, recieved, balance, and number of transactions it has been used in.'
          schema:
            x-oad-type: reference
            $ref: '#/definitions/AddressSummary'
          x-oad-type: response
      parameters:
        -
          $ref: '#/parameters/address'
          x-oad-type: reference
    x-oad-type: operation
  /SQL:
    get:
      operationId: 'SQL Query'
      summary: 'Use SQL in a RESTful way'
      description: 'API exposed for sending SQL queries directly against the chainquery database. Since this is an exposed API there are limits to its use. These limits include queries per hour, read-only, limited to 60 second timeout. '
      tags:
        - query
      responses:
        '200':
          description: 'It will return the rows as a json array.'
          schema:
            x-oad-type: array
            type: array
            items:
              x-oad-type: object
              type: object
              title: 'Row object'
              description: 'object with fields representing the selected columns of the query.'
          x-oad-type: response
        '400':
          description: 'This is returned when the sql query has failed, either due to syntax issues with the query, nameing, priviledges etc. An error is provided with the response.'
          schema:
            x-oad-type: string
            type: string
            title: 'error message'
            description: 'what went wrong?'
          x-oad-type: response
      parameters:
        -
          name: query
          in: path
          description: 'The SQL query to be put against the chainquery database.'
          required: true
          type: string
          x-oad-type: parameter
    x-oad-type: operation
parameters:
  address:
    name: LbryAddress
    in: query
    description: 'A LbryAddress'
    required: true
    type: string
    x-oad-type: parameter
definitions:
  TableSize:
    x-oad-type: object
    type: object
    title: 'Table Size '
    description: 'Contains the name and number of rows for a table.'
    properties:
      TableName:
        x-oad-type: string
        type: string
        title: 'Table Name'
        description: 'Name of the table being referenced.'
      NrRows:
        x-oad-type: integer
        type: integer
        title: 'Number of Rows'
        description: 'The number of rows in the referenced table'
        format: int64
  TableStatus:
    x-oad-type: object
    type: object
    title: 'Table status of Chain Query'
    description: 'Contains an array of tables and their row count.'
    properties:
      Status:
        x-oad-type: array
        type: array
        title: 'Array of Tables'
        items:
          x-oad-type: reference
          $ref: '#/definitions/TableSize'
        uniqueItems: true
  AddressSummary:
    x-oad-type: object
    type: object
    title: 'Contains information about an address'
    properties:
      Address:
        x-oad-type: string
        type: string
        title: 'LBRY Address'
      TotalReceived:
        x-oad-type: number
        type: number
        title: ' Total Recieved'
        description: 'Total amount received by address from all transactions it was a part of.'
        format: double
        default: 0
      TotalSent:
        x-oad-type: number
        type: number
        title: 'Total Sent'
        description: 'Total amount sent from address for all transactions it was a part of.'
        format: double
        default: 0
      Balance:
        x-oad-type: number
        type: number
        title: Balance
        description: 'The current balance of an address'
        format: double
        default: 0
info:
  title: 'Chain Query'
  version: 0.1.0
  description: 'The LBRY blockchain is read into SQL where important structured information can be extracted through the Chain Query API.'
host: ':6300'
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
swagger: '2.0'